+++ nitro/__init__.py
@@ -32,7 +32,7 @@ from nitro.conf import (
     get_setting,
 )
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"
 
 __all__ = [
     # Base